FFmpegExtractor: Don't use our extractor when we agree with StageFright
Prior to this commit, if StageFright picks mimeType X with any
confidence <= .8 then it uses the StageFright extractors and otherwise
(if ffmpeg recognizes the file format) then it will use the ffmpeg
extractors.

This commit changes the logic to prefer the StageFright extractors
(keeping the behaviour closer to aosp and less dependent on ffmpeg's
implementation) whenever both agree on the same mime type.

Change-Id: I98b1b3c3e94756923106911783104ca42e916aed