SoftFFmpegVideo: use AVDISCARD_DEFAULT in skip_loop_filter
After a lot of trials and errors, it appears the rendering problem
has nothing to do with color depth. The problem is that the codec
is configured with skip_loop_filter = AVDISCARD_ALL. I'm not sure
what's the reason for that choice. By comparison, it was using
AVDISCARD_DEFAULT in kitkat-x86 branch. If I change the value to
AVDISCARD_DEFAULT instead, I get correct decoding.

Fixes: a5d3c4c2 "stagefright-plugins: Implement frame dropping and AV sync"