SoftFFmpegVideo: fix adpative video scaling problem
The way I understand it is that the codec is not supposed to scale the
video frames to the buffer size. In case of adpative playback, the frames
may be smaller than the buffer, which corresponds to the max possible
frame size in the adaptive stream.

The patch also drops fiddling with the cropping attributes. I'm not sure
what was the purpose, but I couldn't find equivalent in other sw codecs
like avcdec or mpeg2dec, so I just dropped that and mimiced what the
other sw codecs are doing.