Use new HW accel API from FFMPEG
This removes dependency on fftools source code and allows to migrate to
newer version of FFMPEG. Initialize VAAPI device before opening the
codec, instead of during format negotiation (which leads to memory leak
when multi-threading is enabled, even if thread_count is reset to 1
during get_format).

Validated with VA-API and FFMPEG 3.4.11 (requires updated ffmpeg repo).